requirements.txt
tweepy
newspaper
BeautifulSoup
goose-extractor
nltk
unidecode
sumy
gnp
language_check
langdetect 